fixed rubocop errors
