+++ bin/main.rb
@@ -1,37 +1,39 @@
 require 'nokogiri'
-require_relative '../lib/scraper'
+
 require_relative '../lib/information'
 
-def variables
-  scraper = Scrapper.new
-  name = scraper.take_name
-  price = scraper.take_price
-  review = scraper.take_review
-  url = scraper.take_url
-  printer_list = []
-  name.count.times { printer_list.push([]) }
-  name.count.times { |index| printer_list[index].push(name[index], price[index], review[index], url[index]) }
-  printer_list
+def display
+  puts
+  puts 'This scraper show FDM printers from Bangood store'
+  puts
+  puts 'If you want to sort them by price Press "1"'
+  puts
+  puts 'If you want to sort them by review Press "2"'
+  puts
+  puts 'List will be automatically sort low  to high.'
+  puts
+  print 'Price or Review: '
 end
+display
 
 def show(list)
   variables
   variables.count.times do |i|
-      puts
-      puts "                  PRINTER-#{i+1}"
-      puts
-      puts "  PRINTER NAME :  #{list[i][0]}"
-      puts
-      puts "  PRICE($)     :  #{list[i][1]}"
-      puts
-      puts "  REVIEW       :  #{list[i][2]}"
-      puts 
-      puts "  URL          :  #{list[i][3]}"
-      puts
-      puts "  -----------------------------------------"
+    puts
+    puts "                  PRINTER-#{i + 1}"
+    puts
+    puts "  PRINTER NAME :  #{list[i][0]}"
+    puts
+    puts "  PRICE($)     :  #{list[i][1]}"
+    puts
+    puts "  REVIEW       :  #{list[i][2]}"
+    puts
+    puts "  URL          :  #{list[i][3]}"
+    puts
+    puts '  -----------------------------------------'
   end
- 
 end
+
 def input
   number = gets.chomp.to_i
   puts
@@ -43,11 +45,7 @@ def input
   variables
   result = variables.sort_by { |list| list[number] }
   result.reverse! if letter == 'Y'
-  numbers.include?(number) ?  show(result) : (puts 'you entered wrong number')
+  numbers.include?(number) ? show(result) : (puts 'you entered wrong number')
 end
 
-display
 input
-
-
-
